Refactor #2 - change layout components's file path
