+++ project/dynamic-busan/pages/Notice/script.js
@@ -3,8 +3,8 @@ import { emptyAlarmImage } from '../../src/js/image';
 import { createElement } from '../../src/js/util/dom';
 import Router from '../../src/js/module/RouterWithCB';
 import { moreRight } from '../../src/js/Icon';
-import ListBoard from '../../src/js/ListBoard';
-import PageSlider from '../../src/js/PageSlider';
+import ListBoard from '../../src/js/layout/ListBoard';
+import PageSlider from '../../src/js/layout/PageSlider';
 
 import data from './data.json';
 import './style.css';
