+++ project/dynamic-busan/src/js/layout/ListBoard.js
@@ -1,4 +1,4 @@
-import { createElement } from './util/dom';
+import { createElement } from '../util/dom';
 
 /**
  * @description List Board Element를 랜더할 수 있는 클래스
